Fix Map vertical scrollbar (#91)
Issue : https://github.com/archaiodata/thacer/issues/90
Adjust the height of the Map to match the viewport height and remove
vertical scrollbar

**Description** :    
The problem is caused by a `<p>` in the footer, which adds blank space
at the bottom. Replacing `<p>` with `<span>` solves the problem, and
there is no need to hide overflow of the `<html>` since the map has its
own overflow hidden.
The footer needs further fixing as it conflicts with the attribution
line of the map tiles. The footer line can be removed completely or it
can be added to the map attribution.